% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_kinase.R
\name{build_kinase}
\alias{build_kinase}
\title{build_kinase}
\usage{
build_kinase(data, model)
}
\arguments{
\item{data}{;data_pw_100 object from KRSA}

\item{model}{;data_modeled file from KRSA}
}
\value{
;kinase-class object
}
\description{
Constructs a kinase-class object using a set of defaults.
}
\examples{
TRUE
}
\seealso{
Other ;processing functions: 
\code{\link{name_eval}()}
}
\concept{;processing functions}
