% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_preprocess_kin.R
\name{cov_preprocess_kin}
\alias{cov_preprocess_kin}
\title{cov_preprocess_kin}
\usage{
cov_preprocess_kin(map, peps)
}
\arguments{
\item{map}{;mapping file}

\item{peps}{;list of peptides}
}
\value{
a dataframe detailing percents and ratios
}
\description{
Generates the percentage dataframe from a given list of peptides. The dataframe is presented in
terms of kinases.
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{export_data}()},
\code{\link{generate_total_df_kin}()},
\code{\link{get_LFC}()},
\code{\link{get_counts_kin}()},
\code{\link{global_analysis}()},
\code{\link{global_heatmap}()},
\code{\link{group_compare}()},
\code{\link{make_cov_plot_kin}()},
\code{\link{mega_filter}()},
\code{\link{nonlinear_filter}()}
}
\concept{processing functions}
