% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plot_df_kin.R
\name{generate_plot_df_kin}
\alias{generate_plot_df_kin}
\title{generate_plot_df_kin}
\usage{
generate_plot_df_kin(kinases, data)
}
\arguments{
\item{kinases}{;list of kinases of interest}

\item{data}{;data to pull from}
}
\value{
sorted dataframe
}
\description{
Creates a sorted dataframe used for the coverage plot
}
\examples{
TRUE
}
