% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sub_df_kin.R
\name{generate_sub_df_kin}
\alias{generate_sub_df_kin}
\title{generate_sub_df_kin}
\usage{
generate_sub_df_kin(kinases, percent, ratio)
}
\arguments{
\item{kinases}{;list of kinases}

\item{percent}{;vector of percent values representing the kinase's representation in the subset}

\item{ratio}{;vector of ratios showing the subset counts relative to the total counts}
}
\value{
;subset dataframe for stacking
}
\description{
Creates a dataframe around the subsetted portion of the map.
}
\examples{
TRUE
}
\seealso{
Other ;processing functions: 
\code{\link{build_kinase}()},
\code{\link{name_eval}()}
}
\concept{;processing functions}
