% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_total_df_kin.R
\name{generate_total_df_kin}
\alias{generate_total_df_kin}
\title{generate_total_df_kin}
\usage{
generate_total_df_kin(kinases, percent)
}
\arguments{
\item{kinases}{;list of kinases of intrest}

\item{percent}{;vector of percents that represent the percent coverage in the total fraction}
}
\value{
a dataframe representing the total fraction in the stack
}
\description{
Creates a dataframe around the entirety of the map
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{cov_preprocess_kin}()},
\code{\link{export_data}()},
\code{\link{get_LFC}()},
\code{\link{get_counts_kin}()},
\code{\link{global_analysis}()},
\code{\link{global_heatmap}()},
\code{\link{group_compare}()},
\code{\link{make_cov_plot_kin}()},
\code{\link{mega_filter}()},
\code{\link{nonlinear_filter}()}
}
\concept{processing functions}
