% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts_kin.R
\name{get_counts_kin}
\alias{get_counts_kin}
\title{get_counts_kin}
\usage{
get_counts_kin(df, kinases)
}
\arguments{
\item{df}{;a dataframe to get the counts from}

\item{kinases}{;list of kinases}
}
\value{
;integer vector of kinase counts
}
\description{
Counts the number of times a paticular kianse appears in a mapping file.
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{cov_preprocess_kin}()},
\code{\link{export_data}()},
\code{\link{generate_total_df_kin}()},
\code{\link{get_LFC}()},
\code{\link{global_analysis}()},
\code{\link{global_heatmap}()},
\code{\link{group_compare}()},
\code{\link{make_cov_plot_kin}()},
\code{\link{mega_filter}()},
\code{\link{nonlinear_filter}()}
}
\concept{processing functions}
