% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_heatmap.R
\name{global_heatmap}
\alias{global_heatmap}
\title{global_heatmap}
\usage{
global_heatmap(data, peptides)
}
\arguments{
\item{data}{;one of the dataframes from the KRSA model}

\item{peptides}{;list of peptides}
}
\value{
global heatmap object
}
\description{
Makes a heatmap that shows all the comparisons.
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{cov_preprocess_kin}()},
\code{\link{export_data}()},
\code{\link{generate_total_df_kin}()},
\code{\link{get_LFC}()},
\code{\link{get_counts_kin}()},
\code{\link{global_analysis}()},
\code{\link{group_compare}()},
\code{\link{make_cov_plot_kin}()},
\code{\link{mega_filter}()},
\code{\link{nonlinear_filter}()}
}
\concept{processing functions}
