% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_analysis.R
\name{group_analysis}
\alias{group_analysis}
\title{group_analysis}
\usage{
group_analysis(model, groups, diff_df)
}
\arguments{
\item{model}{;the model object from KRSA}

\item{groups}{;user-specified groups}

\item{diff_df}{;the LFC table generated from the 'export_data()' function}
}
\value{
heatmap object
}
\description{
Generates a peptide list from a group comparison and makes a heatmap
}
\examples{
TRUE
}
