% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_compare.R
\name{group_compare}
\alias{group_compare}
\title{group_compare}
\usage{
group_compare(groups, data, model, threshold)
}
\arguments{
\item{groups}{;user-defined groups}

\item{data}{;data_pw_100 from KRSA}

\item{model}{;the model file from KRSA}

\item{threshold}{;user-defined threshold}
}
\value{
vector of peptides
}
\description{
Generates a peptide list from the indicated groups at the indicated threshold.
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{export_data}()},
\code{\link{get_LFC}()},
\code{\link{global_analysis}()},
\code{\link{mega_filter}()},
\code{\link{nonlinear_filter}()}
}
\concept{processing functions}
