% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_heatmap.R
\name{group_heatmap}
\alias{group_heatmap}
\title{group_heatmap}
\usage{
group_heatmap(data, peptides, groups)
}
\arguments{
\item{data}{;one of the dataframes from the KRSA model object}

\item{peptides}{;list of peptides}

\item{groups}{;user-specified groups}
}
\value{
group heatmap object
}
\description{
Makes a heatmap for a specific group comparison.
}
\examples{
TRUE
}
