% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low_filter.R
\name{low_filter}
\alias{low_filter}
\title{low_filter}
\usage{
low_filter(data, threshold, groups)
}
\arguments{
\item{data}{;data_pw_100 file from KRSA}

\item{threshold}{;signal threshold value}

\item{groups}{;user-defined group comparison}
}
\value{
;filtered dataframe
}
\description{
Filters out peptides with a low signal.
}
\examples{
TRUE
}
