% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cov_plot_kin.R
\name{make_cov_plot_kin}
\alias{make_cov_plot_kin}
\title{make_cov_plot_kin}
\usage{
make_cov_plot_kin(data)
}
\arguments{
\item{data}{;dataframe to use for the plot.}
}
\value{

}
\description{
Makes the coverage plot itself
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{cov_preprocess_kin}()},
\code{\link{export_data}()},
\code{\link{generate_total_df_kin}()},
\code{\link{get_LFC}()},
\code{\link{get_counts_kin}()},
\code{\link{global_analysis}()},
\code{\link{global_heatmap}()},
\code{\link{group_compare}()},
\code{\link{mega_filter}()},
\code{\link{nonlinear_filter}()}
}
\concept{processing functions}
