% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_eval.R
\name{name_eval}
\alias{name_eval}
\title{name_eval}
\usage{
name_eval(data, vec)
}
\arguments{
\item{data}{;a dataframe consisting of the combined SigmBg and SigSat files}

\item{vec}{;a vector of names to replace the original sample names}
}
\value{
;a modified dataframe with the replaced sample names
}
\description{
Sets sample names during the KRSA process to a standard format based on their corresponding array values.
}
\examples{
TRUE
}
\seealso{
Other ;processing functions: 
\code{\link{build_kinase}()},
\code{\link{generate_sub_df_kin}()}
}
\concept{;processing functions}
