% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinear_filter.R
\name{nonlinear_filter}
\alias{nonlinear_filter}
\title{nonlinear_filter}
\usage{
nonlinear_filter(data, threshold, groups)
}
\arguments{
\item{data}{;dataframe from the model file from KRSA}

\item{threshold}{;r-seq threshold}

\item{groups}{;user-defined groups}
}
\value{
a filtered dataframe
}
\description{
Filters out peptides that do not fit the model.
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{cov_preprocess_kin}()},
\code{\link{export_data}()},
\code{\link{generate_total_df_kin}()},
\code{\link{get_LFC}()},
\code{\link{get_counts_kin}()},
\code{\link{global_analysis}()},
\code{\link{global_heatmap}()},
\code{\link{group_compare}()},
\code{\link{make_cov_plot_kin}()},
\code{\link{mega_filter}()}
}
\concept{processing functions}
