% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_filter.R
\name{ref_filter}
\alias{ref_filter}
\title{ref_filter}
\usage{
ref_filter(peptides)
}
\arguments{
\item{peptides}{;a list of peptides}
}
\value{
a vector of filtered peptides
}
\description{
Filters out refrence peptides.
}
\examples{
TRUE
}
\concept{processing function}
