% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swoosh_plot_kin.R
\name{swoosh_plot_kin}
\alias{swoosh_plot_kin}
\title{swoosh_plot_kin}
\usage{
swoosh_plot_kin(kinases, data)
}
\arguments{
\item{kinases}{;list of kinases}

\item{data}{;data frame to work off of}
}
\value{
the coverage plot
}
\description{
Wrapper function that sets up the plot dataframe and creates the plot from it.
}
\examples{
TRUE
}
\seealso{
Other processing function: 
\code{\link{ref_filter}()}
}
\concept{processing function}
