% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_slider.R
\name{thresholdSlider}
\alias{thresholdSlider}
\title{threshold_slider}
\usage{
thresholdSlider(id)
}
\arguments{
\item{id}{;id for the widget}
}
\value{
;a slider for entering a theshold
}
\description{
Sets up the threshold slider that can be interacted with by the user.
}
\examples{
TRUE
}
