% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{export_data}
\usage{
export_data(groups, data, model, threshold)
}
\arguments{
\item{groups}{;user-defined group}

\item{data}{;data_pw_100 from KRSA}

\item{model}{;the model file from KRSA}

\item{threshold}{;user-defined threshold}
}
\value{
data frame of peptide names and LFC values
}
\description{
generates a LFC table for the given group comparison at the specified threshold
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{cov_preprocess_kin}()},
\code{\link{generate_total_df_kin}()},
\code{\link{get_LFC}()},
\code{\link{get_counts_kin}()},
\code{\link{global_analysis}()},
\code{\link{global_heatmap}()},
\code{\link{group_compare}()},
\code{\link{make_cov_plot_kin}()},
\code{\link{mega_filter}()},
\code{\link{nonlinear_filter}()}
}
\concept{processing functions}
