% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_heatmap.R
\name{global_heatmap}
\alias{global_heatmap}
\title{global_heatmap}
\usage{
global_heatmap(data, peptides)
}
\arguments{
\item{data}{;one of the dataframes from the KRSA model}

\item{peptides}{;list of peptides}
}
\value{
global heatmap object
}
\description{
Makes a heatmap that shows all the comparisons.
}
\examples{
TRUE
}
