% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mega_filter.R
\name{mega_filter}
\alias{mega_filter}
\title{mega_filter}
\usage{
mega_filter(data, model, signal_threshold, r2_threshold, groups)
}
\arguments{
\item{data}{;data_pw_100 file from KRSA}

\item{signal_threshold}{;threshold to filter the low signal peptides}

\item{r2_threshold}{;threshold to filter out the nonlinear peptides}

\item{groups}{;user-defined peptides}

\item{data2}{;model file from KRSA}
}
\value{
a vector of pepties
}
\description{
Combines multiple filter functions into one.
}
\examples{
TRUE
}
\seealso{
Other processing functions: 
\code{\link{export_data}()},
\code{\link{get_LFC}()},
\code{\link{global_analysis}()},
\code{\link{group_compare}()},
\code{\link{nonlinear_filter}()}
}
\concept{processing functions}
